# backend/spec/openapi.yaml
openapi: "3.1.0"
info:
  title: InfraMind API
  version: "0.1.0"
  description: Autonomous Infrastructure Intelligence Platform
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /assets:
    get:
      summary: List all assets
      responses:
        "200":
          description: Asset list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      summary: Create an asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      summary: Get asset by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Asset detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      summary: Soft delete an asset
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
...
  /assets/{id}/autonomy:
    get:
      summary: Get asset autonomy mode
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Autonomy mode
          content:
            application/json:
              schema:
                type: object
                properties:
                  autonomyMode: { type: string, enum: [manual, advisory, autonomous] }
    patch:
      summary: Update asset autonomy mode
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [autonomyMode]
              properties:
                autonomyMode: { type: string, enum: [manual, advisory, autonomous] }
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
...
  /devices:
    post:
      summary: Register a device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResponse"
  /devices/{id}:
    get:
      summary: Get device by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Device detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
  /devices/{id}/heartbeat:
    post:
      summary: Update device heartbeat
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Heartbeat updated
  /devices/{id}/config:
    get:
      summary: Get device configuration
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Device configuration
          content:
            application/json:
              schema:
                type: object
    put:
      summary: Update device configuration
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                config:
                  type: object
      responses:
        "200":
          description: Configuration updated
          content:
            application/json:
              schema:
                type: object
...
  /actions:
    get:
      summary: List autonomous actions with optional filters
      parameters:
        - name: asset_id
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { type: string, enum: [proposed, approved, rejected, executed, failed] }
        - name: type
          in: query
          schema: { type: string, enum: [command, restart, config_change, notification] }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: Action list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Action"
    post:
      summary: Propose an autonomous action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposeActionRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Action"
  /actions/{id}:
    get:
      summary: Get action by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Action detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Action"
  /actions/{id}/approve:
    patch:
      summary: Approve an action
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Action"
  /actions/{id}/reject:
    patch:
      summary: Reject an action
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Action"
  /organizations:
    get:
      summary: List organizations (admin only)
      responses:
        "200":
          description: Organization list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      summary: Create an organization (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{id}:
    get:
      summary: Get organization by ID (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Organization detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    patch:
      summary: Update an organization (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
...
components:
  schemas:
    Asset:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        type: { type: string }
        autonomyMode: { type: string, enum: [manual, advisory, autonomous] }
        location: { type: object }
        metadata: { type: object }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateAssetRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        type: { type: string, default: transformer }
        location: { type: object }
        metadata: { type: object }
    Device:
      type: object
      properties:
        id: { type: string, format: uuid }
        assetId: { type: string, format: uuid }
        firmwareVersion: { type: string }
        status: { type: string }
        location: { type: object }
        lastHeartbeat: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    RegisterDeviceRequest:
      type: object
      required: [assetId]
      properties:
        assetId: { type: string, format: uuid }
        firmwareVersion: { type: string }
        location: { type: object }
    RegistrationResponse:
      type: object
      properties:
        device:
          $ref: "#/components/schemas/Device"
        mqttUsername: { type: string }
        mqttPassword: { type: string }
...
    Action:
      type: object
      properties:
        id: { type: string, format: uuid }
        assetId: { type: string, format: uuid }
        deviceId: { type: string, format: uuid, nullable: true }
        type: { type: string, enum: [command, restart, config_change, notification] }
        payload: { type: object }
        source: { type: string, enum: [ai, manual] }
        status: { type: string, enum: [proposed, approved, rejected, executed, failed] }
        approvalRequired: { type: boolean }
        autoApproved: { type: boolean }
        reason: { type: string }
        result: { type: string, nullable: true }
        proposedAt: { type: string, format: date-time }
        executedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProposeActionRequest:
      type: object
      required: [assetId, type]
      properties:
        assetId: { type: string, format: uuid }
        deviceId: { type: string, format: uuid }
        type: { type: string, enum: [command, restart, config_change, notification] }
        payload: { type: object }
        source: { type: string, enum: [ai, manual] }
        reason: { type: string }
        approvalRequired: { type: boolean }
    Organization:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        settings: { type: object }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateOrganizationRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        slug: { type: string }
        settings: { type: object }
    UpdateOrganizationRequest:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        settings: { type: object }
...